[Chore] 슬라이더 프로퍼티 Float으로 수정

+++ RelaxOn/Views/Kitchen/CustomSlider.swift
@@ -10,12 +10,12 @@ import Foundation
 
 struct CustomSlider<Component: View>: View {
 
-    @Binding var value: Double
-    var range: (Double, Double)
+    @Binding var value: Float
+    var range: (Float, Float)
     var knobWidth: CGFloat?
     let viewBuilder: (CustomSliderComponents) -> Component
 
-    init(value: Binding<Double>, range: (Double, Double), knobWidth: CGFloat? = nil,
+    init(value: Binding<Float>, range: (Float, Float), knobWidth: CGFloat? = nil,
          _ viewBuilder: @escaping (CustomSliderComponents) -> Component
     ) {
         _value = value
@@ -25,9 +25,9 @@ struct CustomSlider<Component: View>: View {
     }
     
     private func onDragChange(_ drag: DragGesture.Value,_ frame: CGRect) {
-        let width = (knob: Double(knobWidth ?? frame.size.height), view: Double(frame.size.width))
-        let xrange = (min: Double(0), max: Double(width.view - width.knob))
-        var value = Double(drag.startLocation.x + drag.translation.width) // knob center x
+        let width = (knob: Float(knobWidth ?? frame.size.height), view: Float(frame.size.width))
+        let xrange = (min: Float(0), max: Float(width.view - width.knob))
+        var value = Float(drag.startLocation.x + drag.translation.width) // knob center x
         value -= 0.5*width.knob // offset from center to leading edge of knob
         value = value > xrange.max ? xrange.max : value // limit to leading edge
         value = value < xrange.min ? xrange.min : value // limit to trailing edge
@@ -38,7 +38,7 @@ struct CustomSlider<Component: View>: View {
     
     private func getOffsetX(frame: CGRect) -> CGFloat {
         let width = (knob: knobWidth ?? frame.size.height, view: frame.size.width)
-        let xrange: (Double, Double) = (0, Double(width.view - width.knob))
+        let xrange: (Float, Float) = (0, Float(width.view - width.knob))
         let result = self.value.convert(fromRange: range, toRange: xrange)
         return CGFloat(result)
     }
@@ -93,12 +93,12 @@ struct CustomSliderModifier: ViewModifier {
     }
 }
 
-extension Double {
-    func convert(fromRange: (Double, Double), toRange: (Double, Double)) -> Double {
+extension Float {
+    func convert(fromRange: (Float, Float), toRange: (Float, Float)) -> Float {
         // Example: if self = 1, fromRange = (0,2), toRange = (10,12) -> solution = 11
         var value = self
         value -= fromRange.0
-        value /= Double(fromRange.1 - fromRange.0)
+        value /= Float(fromRange.1 - fromRange.0)
         value *= toRange.1 - toRange.0
         value += toRange.0
         return value
